backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
whitenoise>=6.6.0
gunicorn>=21.2.0 
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
tabulate>=0.9.0